github-profile-analyzer-main/github-profile-analyzer-main/requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
textblob
wordcloud
matplotlib
python-dotenv
nltk
PyGithub
pillow
pyvis
anthropic
